agents: isolate document preview loading policy
