+++ packages/agent-ui/src/document-preview/content.ts
@@ -1,17 +1,7 @@
-import { getAgentMessageFileContent, isSixbApiError } from "@sixb/client"
+import { getAgentMessageFileContent } from "@sixb/client"
 import { useQuery } from "@tanstack/react-query"
+import { documentLoadError, textPreviewTooLarge } from "./content-policy"
 import type { AgentDocumentSource } from "./types"
-
-export const MAX_TEXT_PREVIEW_BYTES = 5 * 1024 * 1024
-export const MAX_MARKDOWN_PREVIEW_BYTES = MAX_TEXT_PREVIEW_BYTES
-
-export function textPreviewTooLarge(source: AgentDocumentSource): boolean {
-  return source.fileRef.sizeBytes > MAX_TEXT_PREVIEW_BYTES
-}
-
-export function markdownPreviewTooLarge(source: AgentDocumentSource): boolean {
-  return textPreviewTooLarge(source)
-}
 
 export function useMarkdownDocument(source: AgentDocumentSource) {
   return useTextDocument(source)
@@ -55,12 +45,4 @@ function useTextDocument(source: AgentDocumentSource) {
     error: query.error ? documentLoadError(query.error) : null,
     tooLarge,
   }
-}
-
-export function documentLoadError(error: unknown): string {
-  if (isSixbApiError(error)) {
-    if (error.status === 404) return "This document is no longer available."
-    if (error.status === 501) return "Document storage is not available."
-  }
-  return "Could not load this document."
 }
